+++ src/main/java/com/secuconnect/client/api/PaymentTransactionsApi.java
@@ -12,7 +12,6 @@ import com.secuconnect.client.model.PaymentTransactionsIncreaseAmountDTO;
 import com.secuconnect.client.model.PaymentTransactionsInstructionsModel;
 import com.secuconnect.client.model.PaymentTransactionsList;
 import com.secuconnect.client.model.PaymentTransactionsProductModel;
-import com.secuconnect.client.model.PaymentTransactionsShippingUrl;
 import com.secuconnect.client.model.ProductExceptionPayload;
 import com.secuconnect.client.model.ResultBoolean;
 import com.secuconnect.client.model.SecupayTransactionProductDTO;
@@ -1011,92 +1010,6 @@ public class PaymentTransactionsApi {
     public Call getPaymentTransactionsOldFormatAsync(String paymentTransactionId, final ApiCallback<SecupayTransactionProductModel> callback) throws ApiException {
         Call call = getPaymentTransactionsOldFormatValidateBeforeCall(paymentTransactionId);
         Type localVarReturnType = new TypeToken<SecupayTransactionProductModel>(){}.getType();
-        apiClient.executeAsync(call, localVarReturnType, callback);
-        return call;
-    }
-
-    /**
-     * Build call for getShippingUrl
-     * @param paymentTransactionId Payment Transaction ID (required)
-     * @return Call to execute
-     * @throws ApiException If fail to serialize the request body object
-     */
-    public Call getShippingUrlCall(String paymentTransactionId) throws ApiException {
-        Object localVarPostBody = null;
-
-        // create path and map variables
-        String localVarPath = "/Payment/Transactions/{paymentTransactionId}/shippingUrl"
-            .replaceAll("\\{" + "paymentTransactionId" + "\\}", apiClient.escapeString(paymentTransactionId.toString()));
-
-        List<Pair> localVarQueryParams = new ArrayList<Pair>();
-        List<Pair> localVarCollectionQueryParams = new ArrayList<Pair>();
-
-        Map<String, String> localVarHeaderParams = new HashMap<String, String>();
-
-        Map<String, Object> localVarFormParams = new HashMap<String, Object>();
-
-        final String[] localVarAccepts = {
-            "application/json"
-        };
-        final String localVarAccept = apiClient.selectHeaderAccept(localVarAccepts);
-        if (localVarAccept != null) localVarHeaderParams.put("Accept", localVarAccept);
-
-        final String[] localVarContentTypes = {
-            
-        };
-        final String localVarContentType = apiClient.selectHeaderContentType(localVarContentTypes);
-        localVarHeaderParams.put("Content-Type", localVarContentType);
-
-        String[] localVarAuthNames = new String[] { "oauth_token" };
-        return apiClient.buildCall(localVarPath, "GET", localVarQueryParams, localVarCollectionQueryParams, localVarPostBody, localVarHeaderParams, localVarFormParams, localVarAuthNames);
-    }
-
-    @SuppressWarnings("rawtypes")
-    private Call getShippingUrlValidateBeforeCall(String paymentTransactionId) throws ApiException {
-        // verify the required parameter 'paymentTransactionId' is set
-        if (paymentTransactionId == null) {
-            throw new ApiException("Missing the required parameter 'paymentTransactionId' when calling getShippingUrl(Async)");
-        }
-
-        return getShippingUrlCall(paymentTransactionId);
-    }
-
-    /**
-     * Shipping URL
-     * Returns the URL to file the shipping information. This works only with invoice payment, and requires the Payment Transaction has status 85, \&quot;Kauf auf Rechnung genehmigt\&quot;.  See also &#x60;PUT /Payment/Transactions/{id}/ShippingInformation&#x60;, _Add shipping information_, to add the data programmatically.
-     * @param paymentTransactionId Payment Transaction ID (required)
-     * @return PaymentTransactionsShippingUrl
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
-     */
-    public PaymentTransactionsShippingUrl getShippingUrl(String paymentTransactionId) throws ApiException {
-        ApiResponse<PaymentTransactionsShippingUrl> resp = getShippingUrlWithHttpInfo(paymentTransactionId);
-        return resp.getData();
-    }
-
-    /**
-     * Shipping URL
-     * Returns the URL to file the shipping information. This works only with invoice payment, and requires the Payment Transaction has status 85, \&quot;Kauf auf Rechnung genehmigt\&quot;.  See also &#x60;PUT /Payment/Transactions/{id}/ShippingInformation&#x60;, _Add shipping information_, to add the data programmatically.
-     * @param paymentTransactionId Payment Transaction ID (required)
-     * @return ApiResponse&lt;PaymentTransactionsShippingUrl&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
-     */
-    public ApiResponse<PaymentTransactionsShippingUrl> getShippingUrlWithHttpInfo(String paymentTransactionId) throws ApiException {
-        Call call = getShippingUrlValidateBeforeCall(paymentTransactionId);
-        Type localVarReturnType = new TypeToken<PaymentTransactionsShippingUrl>(){}.getType();
-        return apiClient.execute(call, localVarReturnType);
-    }
-
-    /**
-     * Shipping URL (asynchronously)
-     * Returns the URL to file the shipping information. This works only with invoice payment, and requires the Payment Transaction has status 85, \&quot;Kauf auf Rechnung genehmigt\&quot;.  See also &#x60;PUT /Payment/Transactions/{id}/ShippingInformation&#x60;, _Add shipping information_, to add the data programmatically.
-     * @param paymentTransactionId Payment Transaction ID (required)
-     * @param callback The callback to be executed when the API call finishes
-     * @return The request call
-     * @throws ApiException If fail to process the API call, e.g. serializing the request body object
-     */
-    public Call getShippingUrlAsync(String paymentTransactionId, final ApiCallback<PaymentTransactionsShippingUrl> callback) throws ApiException {
-        Call call = getShippingUrlValidateBeforeCall(paymentTransactionId);
-        Type localVarReturnType = new TypeToken<PaymentTransactionsShippingUrl>(){}.getType();
         apiClient.executeAsync(call, localVarReturnType, callback);
         return call;
     }